Spring Security demonstration
If we had the `spring-boot-starter-security` starter, our app is secured
using basic auth. This simple behaviour gets you starter with a default
behaviour and invite you to look at the documentation for more options.

The default creates a user name `user` with an auto-generated `password`
that is printed on startup. That's rather annoying and we'd like to use
a defined set of users. One way to do that is to add our own
`GlobalAuthenticationConfigurerAdapter` config that defines an
`AuthenticationManager` with two users: user has the `USER` role and
hero has the `USER` and `HERO` roles.